feedback_generation/requirements.txt
fastapi[standard]
opencv-python
mediapipe
uvicorn
boto3
python-dotenv
